# Fix/Do not use `Res` pointers in SE