--- a/pkg/local_object_storage/blobovnicza/put.go
+++ b/pkg/local_object_storage/blobovnicza/put.go
@@ -47,7 +47,7 @@ func (p *PutPrm) SetMarshaledObject(data []byte) {
 // Returns ErrFull if blobovnicza is filled.
 //
 // Should not be called in read-only configuration.
-func (b *Blobovnicza) Put(prm PutPrm) (*PutRes, error) {
+func (b *Blobovnicza) Put(prm PutPrm) (PutRes, error) {
 	sz := uint64(len(prm.objData))
 	bucketName := bucketForSize(sz)
 	key := addressKey(prm.addr)
@@ -76,7 +76,7 @@ func (b *Blobovnicza) Put(prm PutPrm) (*PutRes, error) {
 		b.incSize(sz)
 	}
 
-	return nil, err
+	return PutRes{}, err
 }
 
 func addressKey(addr oid.Address) []byte {