--- a/pkg/local_object_storage/blobstor/delete_big.go
+++ b/pkg/local_object_storage/blobstor/delete_big.go
@@ -22,7 +22,7 @@ type DeleteBigRes struct{}
 // to completely remove the object.
 //
 // Returns an error of type apistatus.ObjectNotFound if there is no object to delete.
-func (b *BlobStor) DeleteBig(prm DeleteBigPrm) (*DeleteBigRes, error) {
+func (b *BlobStor) DeleteBig(prm DeleteBigPrm) (DeleteBigRes, error) {
 	err := b.fsTree.Delete(prm.addr)
 	if errors.Is(err, fstree.ErrFileNotFound) {
 		var errNotFound apistatus.ObjectNotFound
@@ -34,5 +34,5 @@ func (b *BlobStor) DeleteBig(prm DeleteBigPrm) (*DeleteBigRes, error) {
 		storagelog.Write(b.log, storagelog.AddressField(prm.addr), storagelog.OpField("fstree DELETE"))
 	}
 
-	return nil, err
+	return DeleteBigRes{}, err
 }