--- a/pkg/local_object_storage/blobstor/get_range_big.go
+++ b/pkg/local_object_storage/blobstor/get_range_big.go
@@ -28,38 +28,38 @@ type GetRangeBigRes struct {
 //
 // Returns ErrRangeOutOfBounds if the requested object range is out of bounds.
 // Returns an error of type apistatus.ObjectNotFound if object is missing.
-func (b *BlobStor) GetRangeBig(prm GetRangeBigPrm) (*GetRangeBigRes, error) {
+func (b *BlobStor) GetRangeBig(prm GetRangeBigPrm) (GetRangeBigRes, error) {
 	// get compressed object data
 	data, err := b.fsTree.Get(prm.addr)
 	if err != nil {
 		if errors.Is(err, fstree.ErrFileNotFound) {
 			var errNotFound apistatus.ObjectNotFound
 
-			return nil, errNotFound
+			return GetRangeBigRes{}, errNotFound
 		}
 
-		return nil, fmt.Errorf("could not read object from fs tree: %w", err)
+		return GetRangeBigRes{}, fmt.Errorf("could not read object from fs tree: %w", err)
 	}
 
 	data, err = b.decompressor(data)
 	if err != nil {
-		return nil, fmt.Errorf("could not decompress object data: %w", err)
+		return GetRangeBigRes{}, fmt.Errorf("could not decompress object data: %w", err)
 	}
 
 	// unmarshal the object
 	obj := objectSDK.New()
 	if err := obj.Unmarshal(data); err != nil {
-		return nil, fmt.Errorf("could not unmarshal the object: %w", err)
+		return GetRangeBigRes{}, fmt.Errorf("could not unmarshal the object: %w", err)
 	}
 
 	payload := obj.Payload()
 	ln, off := prm.rng.GetLength(), prm.rng.GetOffset()
 
 	if pLen := uint64(len(payload)); pLen < ln+off {
-		return nil, object.ErrRangeOutOfBounds
+		return GetRangeBigRes{}, object.ErrRangeOutOfBounds
 	}
 
-	return &GetRangeBigRes{
+	return GetRangeBigRes{
 		rangeData: rangeData{
 			data: payload[off : off+ln],
 		},