Based on and blocked by #1418.

Benchmark of the `meta.Get`:

Comparing just `*Res` -> `Res`:
```
name               old time/op    new time/op    delta
Get/1_objects-8      9.69µs ±13%    9.68µs ± 7%    ~     (p=0.954 n=10+10)
Get/10_objects-8      102µs ±10%     100µs ± 4%    ~     (p=0.393 n=10+10)
Get/100_objects-8    1.00ms ±10%    1.03ms ±13%    ~     (p=0.280 n=10+10)

name               old alloc/op   new alloc/op   delta
Get/1_objects-8      3.22kB ± 0%    3.22kB ± 0%  -0.25%  (p=0.000 n=9+9)
Get/10_objects-8     33.5kB ± 1%    33.4kB ± 1%  -0.38%  (p=0.010 n=9+10)
Get/100_objects-8     336kB ± 0%     335kB ± 0%    ~     (p=0.059 n=9+9)

name               old allocs/op  new allocs/op  delta
Get/1_objects-8        55.0 ± 0%      54.0 ± 0%  -1.82%  (p=0.000 n=9+9)
Get/10_objects-8        576 ± 1%       565 ± 1%  -1.88%  (p=0.001 n=10+10)
Get/100_objects-8     5.80k ± 0%     5.70k ± 0%  -1.71%  (p=0.000 n=9+9)
```

Comparing `Get(Prm) *Res` -> `Get(Prm, *Res)`:
```
name               old time/op    new time/op    delta
Get/1_objects-8      9.69µs ±13%    9.87µs ±21%    ~     (p=0.842 n=10+9)
Get/10_objects-8      102µs ±10%      99µs ± 5%    ~     (p=0.211 n=10+9)
Get/100_objects-8    1.00ms ±10%    1.00ms ± 5%    ~     (p=0.579 n=10+10)

name               old alloc/op   new alloc/op   delta
Get/1_objects-8      3.22kB ± 0%    3.22kB ± 0%  -0.25%  (p=0.000 n=9+9)
Get/10_objects-8     33.5kB ± 1%    33.5kB ± 1%    ~     (p=0.170 n=9+10)
Get/100_objects-8     336kB ± 0%     334kB ± 0%  -0.37%  (p=0.000 n=9+9)

name               old allocs/op  new allocs/op  delta
Get/1_objects-8        55.0 ± 0%      54.0 ± 0%  -1.82%  (p=0.000 n=9+9)
Get/10_objects-8        576 ± 1%       570 ± 0%  -1.02%  (p=0.026 n=10+8)
Get/100_objects-8     5.80k ± 0%     5.69k ± 0%  -1.75%  (p=0.000 n=9+9)
```

Seems almost equal but I do not like creating `Res` struct on the caller side and do not see real use cases when that could help much.