@carpawell 2022-05-31
> No improvements according to provided results.

What improvements did you expect? We have got rid of one allocation for one `Get` call. Also, `res` does not escape to the heap so it also should help GC (I do not know how to test that). My bench is kinda trivial, I may rewrite it to get all the objects for every test (for 1/10/100 objects) then it would have more informative results.